+++ src/Symfony/Bundle/TwigBundle/CacheWarmer/TemplateCacheWarmer.php
@@ -37,6 +37,8 @@ class TemplateCacheWarmer implements CacheWarmerInterface, ServiceSubscriberInte
 
     /**
      * {@inheritdoc}
+     *
+     * @return string[] A list of template files to preload on PHP 7.4+
      */
     public function warmUp(string $cacheDir)
     {
@@ -44,14 +46,22 @@ class TemplateCacheWarmer implements CacheWarmerInterface, ServiceSubscriberInte
             $this->twig = $this->container->get('twig');
         }
 
+        $files = [];
+
         foreach ($this->iterator as $template) {
             try {
-                $this->twig->load($template);
+                $template = $this->twig->load($template);
+
+                if (\is_callable([$template, 'unwrap'])) {
+                    $files[] = (new \ReflectionClass($template->unwrap()))->getFileName();
+                }
             } catch (Error $e) {
                 // problem during compilation, give up
                 // might be a syntax error or a non-Twig template
             }
         }
+
+        return $files;
     }
 
     /**