+++ src/Symfony/Component/DependencyInjection/Dumper/Preloader.php
@@ -13,12 +13,31 @@ namespace Symfony\Component\DependencyInjection\Dumper;
 
 /**
  * @author Nicolas Grekas <p@tchwork.com>
- *
- * @internal
  */
-class Preloader
+final class Preloader
 {
-    public static function preload(array $classes)
+    public static function append(string $file, array $list): void
+    {
+        if (!file_exists($file)) {
+            throw new \LogicException(sprintf('File "%s" does not exist.', $file));
+        }
+
+        $cacheDir = \dirname($file);
+        $classes = [];
+
+        foreach ($list as $item) {
+            if (0 === strpos($item, $cacheDir)) {
+                file_put_contents($file, sprintf("require __DIR__.%s;\n", var_export(substr($item, \strlen($cacheDir)), true)), FILE_APPEND);
+                continue;
+            }
+
+            $classes[] = sprintf("\$classes[] = %s;\n", var_export($item, true));
+        }
+
+        file_put_contents($file, sprintf("\n\$classes = [];\n%sPreloader::preload(\$classes);\n", implode('', $classes)), FILE_APPEND);
+    }
+
+    public static function preload(array $classes): void
     {
         set_error_handler(function ($t, $m, $f, $l) {
             if (error_reporting() & $t) {